changed code as per latest changes on main branch of hotwax and improved code quality (#1x634cr).
