+++ .env.example
@@ -1,6 +1,5 @@
 VUE_APP_I18N_LOCALE=en
 VUE_APP_I18N_FALLBACK_LOCALE=en
 VUE_APP_CACHE_MAX_AGE=3600
-VUE_APP_BASE_URL=https://dev-hc.hotwax.io/api/
 VUE_APP_VIEW_SIZE=10
 VUE_APP_PRDT_IDENT_TYPE_ID=SHOPIFY_PROD_SKU
