+++ src/locales/en.json
@@ -1,13 +1,13 @@
 {
-  "ATP": "ATP",
   "Add": "Add",
   "Adjust quantity": "Adjust quantity",
   "All orders": "All orders",
   "Allocations": "Allocations",
   "Arrival date": "Arrival date",
+  "arrival date": "arrival date",
+  "ATP": "ATP",
   "Auto cancel": "Auto cancel",
   "Back orders": "Back orders",
-  "Blank": "Blank",
   "Catalog": "Catalog",
   "Categories": "Categories",
   "Change fulfillment location": "Change fulfillment location",
@@ -15,6 +15,7 @@
   "Color": "Color",
   "Current QOH": "Current QOH",
   "Date": "Date",
+  "delivery": "delivery",
   "Destination": "Destination",
   "Download orders": "Download orders",
   "Edit address": "Edit address",
@@ -27,61 +28,66 @@
   "Edit quantity on hand": "Edit quantity on hand",
   "Edit safety stock": "Edit safety stock",
   "Edit threshold": "Edit threshold",
-  "Email": "Email",
   "Estimated arrival": "Estimated arrival",
   "Expedited": "Expedited",
   "Facility": "Facility",
+  "Fulfillment": "Fulfillment",
   "Fulfillment options": "Fulfillment options",
   "Fulfillment settings": "Fulfillment settings",
-  "Fulfillment": "Fulfillment",
-  "Full Name": "Full Name",
   "General information": "General information",
   "Group by": "Group by",
   "In Stock": "{count} In Stock",
   "In stock": "In stock",
   "Internal ID": "Internal ID",
   "Items": "Items",
+  "items": "items",
   "Last brokered": "Last brokered",
+  "Location": "Location",
   "Location Inventory": "Location Inventory",
   "Location inventory": "Location inventory",
-  "Location": "Location",
   "Locations": "Locations",
+  "locations": "locations",
   "Logout": "Logout",
   "Logs": "Logs",
   "OMS": "OMS",
   "On order": "On order",
   "Open orders": "Open orders",
   "Options": "Options",
+  "Order": "Order",
+  "order created": "order created",
+  "Order detail": "Order detail",
   "Order ID": "Order ID",
   "Order Name": "Order Name",
   "Order Number": "Order Number",
-  "Order detail": "Order detail",
-  "Order": "Order",
+  "ordered": "ordered",
   "Ordered on": "Ordered on",
   "Orders": "Orders",
+  "orders": "orders",
   "Overwrite": "Overwrite",
+  "pickup": "pickup",
   "PO arrival date": "PO arrival date",
-  "Phone Number": "Phone Number",
   "Pre orders": "Pre orders",
   "Pre-order": "Pre-order",
-  "Product Store": "Product Store",
-  "Product inventory detail": "Product inventory detail",
   "Product inventory": "Product inventory",
+  "Product inventory detail": "Product inventory detail",
   "Product name": "Product name",
+  "Product Store": "Product Store",
   "Products": "Products",
   "Promise date": "Promise date",
-  "Purchase Order": "Purchase Order",
   "Purchase date": "Purchase date",
-  "Purchase order detail": "Purchase order detail",
   "Purchase order": "Purchase order",
+  "purchase order ATP": "purchase order ATP",
+  "Purchase order detail": "Purchase order detail",
   "Purchase orders": "Purchase orders",
   "QOH": "QOH",
   "QTY": "QTY",
   "Ready to create an app?": "Ready to create an app?",
   "Receive": "Receive",
+  "received": "received",
+  "reduce inventory": "reduce inventory",
   "Remove": "Remove",
   "Retail": "Retail",
-  "SKU": "SKU",
+  "safety stock": "safety stock",
   "Search purchase orders and products": "Search purchase orders and products",
   "Settings": "Settings",
   "Shipping from": "Shipping from",
@@ -90,37 +96,24 @@
   "Shopify IDs": "Shopify IDs",
   "Show order items": "Show order items",
   "Size": "Size",
-  "Sort by": "Sort by",
+  "SKU": "SKU",
   "Sort": "Sort",
+  "Sort by": "Sort by",
   "Standard": "Standard",
   "Start with Ionic": "Start with Ionic",
-  "Store pickup": "Store pickup",
   "Store": "Store",
+  "store name": "store name",
+  "Store pickup": "Store pickup",
+  "threshold consumed": "threshold consumed",
   "Timeline": "Timeline",
   "Total": "Total",
   "Tracking number": "Tracking number",
   "Type": "Type",
-  "UI Components": "UI Components",
-  "UPC": "UPC",
   "Unfillable": "Unfillable",
+  "UPC": "UPC",
   "Variant": "Variant",
   "View location details": "View location details",
   "View orders": "View orders",
   "Warehouse": "Warehouse",
-  "Without promise date": "Without promise date",
-  "arrival date": "arrival date",
-  "delivery": "delivery",
-  "items": "items",
-  "locations": "locations",
-  "order created": "order created",
-  "ordered": "ordered",
-  "orders": "orders",
-  "pickup": "pickup",
-  "purchase order ATP": "purchase order ATP",
-  "received": "received",
-  "reduce inventory": "reduce inventory",
-  "safety stock": "safety stock",
-  "store name": "store name",
-  "threshold consumed": "threshold consumed",
-  "variants": "variants"
+  "Without promise date": "Without promise date"
 }
